Separate functions into React components
Created separate JavaScript files for React components. Added styling
