+++ src/App.js
@@ -1,0 +1,13 @@
+import Header from './components/Header';
+import Main from './components/Main';
+
+function App() {
+   return (
+       <main>
+           <Header />
+           <Main />
+       </main>
+   );
+ }
+
+ export default App;
