+++ src/components/Main.js
@@ -1,0 +1,18 @@
+function Main() {
+   return (
+       <section>
+            <div className="container">
+                <h1>Fun Facts about React</h1>
+                <ul>
+                    <li>Was first released in 2013</li>
+                    <li>Was originally created by Jordan Walke</li>
+                    <li>Has well over 100K stars on GitHub</li>
+                    <li>Is maintained by Facebook</li>
+                    <li>Powers thousands of enterprise apps, including mobile apps</li>
+                </ul>
+           </div>
+       </section>
+   );
+ }
+
+ export default Main;
